Refactor And Reorganize Types
