+++ packages/bodiless-page/src/MenuOptions/index.ts
@@ -1,3 +1,2 @@
-export * from './types';
 export * from './usePageMenuOptions';
 export * from './withPageMenuOptions';
